[#44853] renamed file-link-list to storage
- file-link-list component renamed to storage component, which is
  container of the list, the header, the info box, and the action bar
- renaming of directories and module
