+++ frontend/src/app/features/work-packages/components/wp-single-view-tabs/files-tab/op-files-tab.html
@@ -24,21 +24,9 @@
     data-qa-selector="op-tab-content--tab-section"
     class="op-tab-content--tab-section"
   >
-    <div class=op-tab-content--header>
-      <span class="op-tab-content--header-icon spot-icon spot-icon_nextcloud-circle op-files-tab--icon op-files-tab--icon_nextcloud"></span>
-      <span class="op-tab-content--header-text" [textContent]="storage.name"></span>
-      <button
-        class="op-tab-content--header-action spot-link"
-        [title]="text.storages.open(storage._links.type.href)"
-        (click)="openStorage(storage._links.open.href)"
-      >
-        <span class="spot-icon spot-icon_external-link"></span>
-      </button>
-    </div>
-
-    <op-file-link-list
+    <op-storage
       [resource]="workPackage"
       [storage]="storage"
-    ></op-file-link-list>
+    ></op-storage>
   </section>
 </div>
