+++ frontend/src/app/shared/components/storages/file-link-icons/file-icons.helper.ts
