+++ frontend/src/app/shared/components/storages/file-link-icons/icon-mappings.ts
