+++ frontend/src/app/shared/components/storages/file-picker-modal/file-picker-modal.html
