+++ frontend/src/app/shared/components/storages/storage-icons/get-icon-for-storage-type.ts
@@ -26,18 +26,12 @@
 // See COPYRIGHT and LICENSE files for more details.
 //++
 
-import {
-  ChangeDetectionStrategy, Component, Input,
-} from '@angular/core';
-import {
-  StorageInformationBox,
-} from 'core-app/shared/components/file-links/storage-information/storage-information-box';
+import { storageIconMappings } from 'core-app/shared/components/storages/storage-icons/icon-mappings';
 
-@Component({
-  selector: 'op-storage-information',
-  templateUrl: './storage-information.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class StorageInformationComponent {
-  @Input() viewModel:StorageInformationBox;
+export default function getIconForStorageType(storageType?:string):string {
+  if (storageType && storageIconMappings[storageType]) {
+    return storageIconMappings[storageType];
+  }
+
+  return storageIconMappings.default;
 }
