+++ frontend/src/app/shared/components/storages/storage-information/storage-action-button.ts
