+++ frontend/src/app/shared/components/storages/storage-information/storage-information.html
