+++ frontend/src/app/shared/components/storages/storage/storage.component.html
@@ -1,6 +1,18 @@
 <ng-container
   *ngIf="resource"
 >
+  <div class=op-tab-content--header>
+    <span class="op-tab-content--header-icon spot-icon spot-icon_nextcloud-circle op-files-tab--icon op-files-tab--icon_nextcloud"></span>
+    <span class="op-tab-content--header-text" [textContent]="storage.name"></span>
+    <button
+      class="op-tab-content--header-action spot-link"
+      [title]="text.openStorage()"
+      (click)="openStorageLocation()"
+    >
+      <span class="spot-icon spot-icon_external-link"></span>
+    </button>
+  </div>
+
   <op-storage-information
     *ngFor="let infoBox of storageInformation | async"
     class="op-files-tab--storage-info-box"
