+++ frontend/src/app/shared/components/storages/storage/storage.component.ts
@@ -50,28 +50,28 @@ import {
   storageAuthorizationError,
   storageConnected,
   storageFailedAuthorization,
-} from 'core-app/shared/components/file-links/file-links-constants.const';
+} from 'core-app/shared/components/storages/storages-constants.const';
 import { CurrentUserService } from 'core-app/core/current-user/current-user.service';
 import { UntilDestroyedMixin } from 'core-app/shared/helpers/angular/until-destroyed.mixin';
 import { I18nService } from 'core-app/core/i18n/i18n.service';
 import isNewResource from 'core-app/features/hal/helpers/is-new-resource';
-import { StorageActionButton } from 'core-app/shared/components/file-links/storage-information/storage-action-button';
+import { StorageActionButton } from 'core-app/shared/components/storages/storage-information/storage-action-button';
 import {
   StorageInformationBox,
-} from 'core-app/shared/components/file-links/storage-information/storage-information-box';
+} from 'core-app/shared/components/storages/storage-information/storage-information-box';
 import { OpModalService } from 'core-app/shared/components/modal/modal.service';
 import { ConfigurationService } from 'core-app/core/config/configuration.service';
 import {
   FilePickerModalComponent,
-} from 'core-app/shared/components/file-links/file-picker-modal/file-picker-modal.component';
+} from 'core-app/shared/components/storages/file-picker-modal/file-picker-modal.component';
 import { IHalResourceLink } from 'core-app/core/state/hal-resource';
 
 @Component({
-  selector: 'op-file-link-list',
-  templateUrl: './file-link-list.html',
+  selector: 'op-storage',
+  templateUrl: './storage.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class FileLinkListComponent extends UntilDestroyedMixin implements OnInit {
+export class StorageComponent extends UntilDestroyedMixin implements OnInit {
   @Input() public resource:HalResource;
 
   @Input() public storage:IStorage;
@@ -108,6 +108,7 @@ export class FileLinkListComponent extends UntilDestroyedMixin implements OnInit
     actions: {
       linkExisting: this.i18n.t('js.storages.link_existing_files'),
     },
+    openStorage: ():string => this.i18n.t('js.storages.open_storage', { storageType: this.storageType }),
   };
 
   public get storageFileLinkingEnabled():boolean {
@@ -160,6 +161,10 @@ export class FileLinkListComponent extends UntilDestroyedMixin implements OnInit
 
   public removeFileLink(fileLink:IFileLink):void {
     this.fileLinkResourceService.remove(this.collectionKey, fileLink);
+  }
+
+  public openStorageLocation():void {
+    window.open(this.storageFilesLocation, '_blank');
   }
 
   public openLinkFilesDialog():void {
@@ -217,7 +222,7 @@ export class FileLinkListComponent extends UntilDestroyedMixin implements OnInit
           if (this.storage._links.authorize) {
             const nonce = uuidv4();
             this.setAuthorizationCallbackCookie(nonce);
-            window.location.href = FileLinkListComponent.authorizationFailureActionUrl(
+            window.location.href = StorageComponent.authorizationFailureActionUrl(
               this.storage._links.authorize.href,
               nonce,
             );
