+++ frontend/src/app/shared/components/storages/storages-constants.const.ts
