+++ frontend/src/app/spot/styles/sass/components/list.sass
@@ -47,7 +47,7 @@
         color: $spot-color-basic-gray-3
         background-color: $spot-color-basic-gray-6
 
-      &_link
+      &_link:not(&_disabled)
         color: var(--content-link-color)
 
     &-floating-wrapper
