+++ frontend/src/global_styles/content/work_packages/shared/_file_list.sass
@@ -33,7 +33,7 @@
     &-action
       text-decoration: none
 
-      &:hover
+      &:not(&_disabled):hover
         .op-file-list--item-title
           text-decoration: underline
